+++ pressed/src/main/scala/de/welt/contentapi/pressed/models/PressedFormats.scala
@@ -17,6 +17,7 @@ object PressedReads {
   implicit lazy val apiCommercial3rdPartyConfigurationReads: Reads[ApiCommercial3rdPartyConfiguration] = Json.reads[ApiCommercial3rdPartyConfiguration]
   implicit lazy val apiCommercialConfigurationReads: Reads[ApiCommercialConfiguration] = Json.reads[ApiCommercialConfiguration]
   implicit lazy val apiThemeConfigurationReads: Reads[ApiThemeConfiguration] = Json.reads[ApiThemeConfiguration]
+  implicit lazy val apiTrackingReportFallbackConfigReads: Reads[ApiTrackingReportFallbackConfiguration] = Json.reads[ApiTrackingReportFallbackConfiguration]
   implicit lazy val apiHeaderConfigurationReads: Reads[ApiHeaderConfiguration] = Json.reads[ApiHeaderConfiguration]
   implicit lazy val apiBrandingConfigurationReads: Reads[ApiSponsoringConfiguration] = Json.reads[ApiSponsoringConfiguration]
   implicit lazy val apiSiteBuildingConfigurationReads: Reads[ApiSiteBuildingConfiguration] = Json.reads[ApiSiteBuildingConfiguration]
@@ -61,6 +62,7 @@ object PressedWrites {
   implicit lazy val apiCommercial3rdPartyConfigurationWrites: Writes[ApiCommercial3rdPartyConfiguration] = Json.writes[ApiCommercial3rdPartyConfiguration]
   implicit lazy val apiCommercialConfigurationWrites: Writes[ApiCommercialConfiguration] = Json.writes[ApiCommercialConfiguration]
   implicit lazy val apiThemeConfigurationWrites: Writes[ApiThemeConfiguration] = Json.writes[ApiThemeConfiguration]
+  implicit lazy val apiTrackingReportFallbackConfigWrites: Writes[ApiTrackingReportFallbackConfiguration] = Json.writes[ApiTrackingReportFallbackConfiguration]
   implicit lazy val apiHeaderConfigurationWrites: Writes[ApiHeaderConfiguration] = Json.writes[ApiHeaderConfiguration]
   implicit lazy val apiBrandingConfigurationWrites: Writes[ApiSponsoringConfiguration] = Json.writes[ApiSponsoringConfiguration]
   implicit lazy val apiSiteBuildingConfigurationWrites: Writes[ApiSiteBuildingConfiguration] = Json.writes[ApiSiteBuildingConfiguration]
@@ -107,6 +109,9 @@ object PressedFormats {
 
   implicit lazy val apiThemeConfigurationFormat: Format[ApiThemeConfiguration] =
     Format(apiThemeConfigurationReads, apiThemeConfigurationWrites)
+
+  implicit lazy val apiTrackingReportFallbackConfigFormat: Format[ApiTrackingReportFallbackConfiguration] =
+    Format(apiTrackingReportFallbackConfigReads, apiTrackingReportFallbackConfigWrites)
 
   implicit lazy val apiHeaderConfigurationFormat: Format[ApiHeaderConfiguration] =
     Format(apiHeaderConfigurationReads, apiHeaderConfigurationWrites)